working on adding pages and menu
