+++ package.json
@@ -11,6 +11,7 @@
     "ethers": "^5.4.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "use-nft": "^0.10.1",
     "web-vitals": "^1.0.1",
@@ -41,13 +42,13 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^9",
+    "autoprefixer": "^10.3.4",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.25.1",
-    "postcss": "^7",
+    "postcss": "^8.3.6",
     "prettier": "^2.3.2",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat"
+    "tailwindcss": "^2.2.9"
   }
 }
