+++ src/components/Profile/Profile.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Profile = (props) => {
+
+    return(
+        <div>Profile</div>
+    )
+}
+  
+export default Profile;
+    
+
+
